requirements.txt
astroid==2.8.4
certifi==2021.10.8
charset-normalizer==2.0.7
idna==3.3
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
paho-mqtt==1.6.1
platformdirs==2.4.0
pylint==2.11.1
requests==2.26.0
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
wrapt==1.13.3
